+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Allow Package XML Viewer to also edit package
 - Add synchronization of JSON Schema in build script
 - Display package and flow contextual menu commands only on relevant files